graph, server: Move data handling for indexing status to graph crate
