+++ server/index-node/src/resolver.rs
@@ -1,10 +1,9 @@
 use graphql_parser::{query as q, schema as s};
 use std::collections::HashMap;
 
-use graph::data::graphql::{
-    object, IntoValue, ObjectOrInterface, TryFromValue, ValueList, ValueMap,
-};
-use graph::data::subgraph::schema::{SubgraphError, SubgraphHealth, SUBGRAPHS_ID};
+use graph::data::graphql::{object, ObjectOrInterface, ValueList, ValueMap};
+use graph::data::subgraph::schema::SUBGRAPHS_ID;
+use graph::data::subgraph::status::Statuses as IndexingStatuses;
 use graph::prelude::*;
 use graph_graphql::prelude::{ExecutionContext, Resolver};
 use std::convert::TryInto;
@@ -50,263 +49,6 @@ pub struct IndexNodeResolver<R, S> {
     logger: Logger,
     graphql_runner: Arc<R>,
     store: Arc<S>,
-}
-
-/// The ID of a subgraph deployment assignment.
-#[derive(Debug)]
-struct DeploymentAssignment {
-    /// ID of the subgraph.
-    subgraph: String,
-    /// ID of the Graph Node that indexes the subgraph.
-    node: String,
-}
-
-impl TryFromValue for DeploymentAssignment {
-    fn try_from_value(value: &q::Value) -> Result<Self, Error> {
-        Ok(Self {
-            subgraph: value.get_required("id")?,
-            node: value.get_required("nodeId")?,
-        })
-    }
-}
-
-/// Light wrapper around `EthereumBlockPointer` that is compatible with GraphQL values.
-#[derive(Debug)]
-struct EthereumBlock(EthereumBlockPointer);
-
-impl From<EthereumBlock> for q::Value {
-    fn from(block: EthereumBlock) -> Self {
-        object! {
-            __typename: "EthereumBlock",
-            hash: block.0.hash_hex(),
-            number: format!("{}", block.0.number),
-        }
-    }
-}
-
-impl IntoValue for EthereumBlock {
-    fn into_value(self) -> q::Value {
-        self.into()
-    }
-}
-
-/// The indexing status of a subgraph on an Ethereum network (like mainnet or ropsten).
-#[derive(Debug)]
-struct EthereumIndexingStatus {
-    /// The network name (e.g. `mainnet`, `ropsten`, `rinkeby`, `kovan` or `goerli`).
-    network: String,
-    /// The current head block of the chain.
-    chain_head_block: Option<EthereumBlock>,
-    /// The earliest block available for this subgraph.
-    earliest_block: Option<EthereumBlock>,
-    /// The latest block that the subgraph has synced to.
-    latest_block: Option<EthereumBlock>,
-}
-
-/// Indexing status information for different chains (only Ethereum right now).
-#[derive(Debug)]
-enum ChainIndexingStatus {
-    Ethereum(EthereumIndexingStatus),
-}
-
-impl From<ChainIndexingStatus> for q::Value {
-    fn from(status: ChainIndexingStatus) -> Self {
-        match status {
-            ChainIndexingStatus::Ethereum(inner) => object! {
-                // `__typename` is needed for the `ChainIndexingStatus` interface
-                // in GraphQL to work.
-                __typename: "EthereumIndexingStatus",
-                network: inner.network,
-                chainHeadBlock: inner.chain_head_block,
-                earliestBlock: inner.earliest_block,
-                latestBlock: inner.latest_block,
-            },
-        }
-    }
-}
-
-/// The overall indexing status of a subgraph.
-#[derive(Debug)]
-struct IndexingStatusWithoutNode {
-    /// The subgraph ID.
-    subgraph: String,
-
-    /// Whether or not the subgraph has synced all the way to the current chain head.
-    synced: bool,
-    health: SubgraphHealth,
-    fatal_error: Option<SubgraphError>,
-    non_fatal_errors: Vec<SubgraphError>,
-
-    /// Indexing status on different chains involved in the subgraph's data sources.
-    chains: Vec<ChainIndexingStatus>,
-}
-
-#[derive(Debug)]
-struct IndexingStatus {
-    /// The subgraph ID.
-    subgraph: String,
-
-    /// Whether or not the subgraph has synced all the way to the current chain head.
-    synced: bool,
-    health: SubgraphHealth,
-    fatal_error: Option<SubgraphError>,
-    non_fatal_errors: Vec<SubgraphError>,
-
-    /// Indexing status on different chains involved in the subgraph's data sources.
-    chains: Vec<ChainIndexingStatus>,
-
-    /// ID of the Graph Node that the subgraph is indexed by.
-    node: String,
-}
-
-impl IndexingStatusWithoutNode {
-    /// Adds a Graph Node ID to the indexing status.
-    fn with_node(self, node: String) -> IndexingStatus {
-        IndexingStatus {
-            subgraph: self.subgraph,
-            synced: self.synced,
-            health: self.health,
-            fatal_error: self.fatal_error,
-            non_fatal_errors: self.non_fatal_errors,
-            chains: self.chains,
-            node,
-        }
-    }
-
-    /// Attempts to parse `${prefix}Hash` and `${prefix}Number` fields on a
-    /// GraphQL object value into an `EthereumBlock`.
-    fn block_from_value(
-        value: &q::Value,
-        prefix: &'static str,
-    ) -> Result<Option<EthereumBlock>, Error> {
-        let hash_key = format!("{}Hash", prefix);
-        let number_key = format!("{}Number", prefix);
-
-        match (
-            value.get_optional::<H256>(hash_key.as_ref())?,
-            value
-                .get_optional::<BigInt>(number_key.as_ref())?
-                .map(|n| n.to_u64()),
-        ) {
-            // Only return an Ethereum block if we can parse both the block hash and number
-            (Some(hash), Some(number)) => {
-                Ok(Some(EthereumBlock(EthereumBlockPointer { hash, number })))
-            }
-            _ => Ok(None),
-        }
-    }
-}
-
-impl TryFromValue for IndexingStatusWithoutNode {
-    fn try_from_value(value: &q::Value) -> Result<Self, Error> {
-        Ok(Self {
-            subgraph: value.get_required("id")?,
-            synced: value.get_required("synced")?,
-            health: value.get_required("health")?,
-            fatal_error: value.get_optional("fatalError")?,
-            non_fatal_errors: value.get_required("nonFatalErrors")?,
-            chains: vec![ChainIndexingStatus::Ethereum(EthereumIndexingStatus {
-                network: value
-                    .get_required::<q::Value>("manifest")?
-                    .get_required::<q::Value>("dataSources")?
-                    .get_values::<q::Value>()?[0]
-                    .get_required("network")?,
-                chain_head_block: Self::block_from_value(value, "ethereumHeadBlock")?,
-                earliest_block: Self::block_from_value(value, "earliestEthereumBlock")?,
-                latest_block: Self::block_from_value(value, "latestEthereumBlock")?,
-            })],
-        })
-    }
-}
-
-impl From<IndexingStatus> for q::Value {
-    fn from(status: IndexingStatus) -> Self {
-        let IndexingStatus {
-            subgraph,
-            chains,
-            fatal_error,
-            health,
-            node,
-            non_fatal_errors,
-            synced,
-        } = status;
-
-        fn subgraph_error_to_value(subgraph_error: SubgraphError) -> q::Value {
-            let SubgraphError {
-                subgraph_id,
-                message,
-                block_ptr,
-                handler,
-                deterministic,
-            } = subgraph_error;
-
-            object! {
-                __typename: "SubgraphError",
-                subgraphId: subgraph_id.to_string(),
-                message: message,
-                handler: handler,
-                block: object! {
-                    __typename: "Block",
-                    number: block_ptr.map(|x| x.number),
-                    hash: block_ptr.map(|x| q::Value::from(Value::Bytes(x.hash.as_ref().into()))),
-                },
-                deterministic: deterministic,
-            }
-        }
-
-        let non_fatal_errors: Vec<q::Value> = non_fatal_errors
-            .into_iter()
-            .map(subgraph_error_to_value)
-            .collect();
-        let fatal_error_val = fatal_error.map_or(q::Value::Null, subgraph_error_to_value);
-
-        object! {
-            __typename: "SubgraphIndexingStatus",
-            subgraph: subgraph,
-            synced: synced,
-            health: q::Value::from(health),
-            fatalError: fatal_error_val,
-            nonFatalErrors: non_fatal_errors,
-            chains: chains.into_iter().map(q::Value::from).collect::<Vec<_>>(),
-            node: node,
-        }
-    }
-}
-
-struct IndexingStatuses(Vec<IndexingStatus>);
-
-impl From<q::Value> for IndexingStatuses {
-    fn from(data: q::Value) -> Self {
-        // Extract deployment assignment IDs from the query result
-        let assignments = data
-            .get_required::<q::Value>("subgraphDeploymentAssignments")
-            .expect("no subgraph deployment assignments in the result")
-            .get_values::<DeploymentAssignment>()
-            .expect("failed to parse subgraph deployment assignments");
-
-        IndexingStatuses(
-            // Parse indexing statuses from deployments
-            data.get_required::<q::Value>("subgraphDeployments")
-                .expect("no subgraph deployments in the result")
-                .get_values()
-                .expect("failed to parse subgraph deployments")
-                .into_iter()
-                // Filter out those deployments for which there is no active assignment
-                .filter_map(|status: IndexingStatusWithoutNode| {
-                    assignments
-                        .iter()
-                        .find(|assignment| assignment.subgraph == status.subgraph)
-                        .map(|assignment| status.with_node(assignment.node.clone()))
-                })
-                .collect(),
-        )
-    }
-}
-
-impl From<IndexingStatuses> for q::Value {
-    fn from(statuses: IndexingStatuses) -> Self {
-        q::Value::List(statuses.0.into_iter().map(q::Value::from).collect())
-    }
 }
 
 impl<R, S> IndexNodeResolver<R, S>
@@ -705,7 +447,7 @@ where
         );
 
         Ok(IndexingStatuses::from(transformed_data)
-            .0
+            .to_vec()
             .into_iter()
             .next()
             .map(Into::into)
